structure: move lib to src
