+++ src/index.js
@@ -9,6 +9,6 @@ String.prototype.includes || (String.prototype.includes = function() {
 });
 
 module.exports={
-	client:require("./lib/client"),
-	Client:require("./lib/client")
+	client:require("./client"),
+	Client:require("./client")
 };
