+++ src/timer.js
